Changed int to long int

+++ src/libasmio.h
@@ -1,5 +1,5 @@
 #ifndef LIBASMIO_H
 #define LIBASMIO_H
-void asmio_gets(const char *buf, int len);
+void asmio_gets(const char *buf, long int len);
 void asmio_printf(const char *format, ...);
 #endif
